Move parser to differnt standalone file
From now on all Parser methods can be reached like:
const parser = require("./parser/Parser");
parser.*name of the function*
